Remove messages file
